% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as.mids.glemb}
\alias{as.mids.glemb}
\title{Convert a glemb object to a mids object}
\usage{
as.mids.glemb(x, data, ...)
}
\arguments{
\item{x}{A \code{glemb} object returned by \code{\link[=glemb]{glemb()}}.}

\item{data}{The original incomplete data frame that was passed to \code{\link[=glemb]{glemb()}}.}

\item{...}{Ignored.}
}
\value{
A \code{mids} object.
}
\description{
Converts a \code{glemb} object to the \code{mids} format used by the \code{mice} package,
enabling use of \code{\link[mice:pool]{mice::pool()}} for results pooling.
}
\details{
The function stacks the original data (as imputation 0) and all \code{m}
imputed datasets in long format, adds \code{.imp} and \code{.id} columns, and
calls \code{\link[mice:as.mids]{mice::as.mids()}} to construct the \code{mids} object.
}
\seealso{
\code{\link[mice:as.mids]{mice::as.mids()}}, \code{\link[mice:pool]{mice::pool()}}
}
