% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.glemb}
\alias{print.glemb}
\title{Print a glemb object}
\usage{
\method{print}{glemb}(x, digits = 3L, ...)
}
\arguments{
\item{x}{A \code{glemb} object returned by \code{\link[=glemb]{glemb()}}.}

\item{digits}{Integer. Number of decimal places for numeric values.
Default is \code{3}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a glemb object
}
