% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.glemb}
\alias{summary.glemb}
\title{Summarise a glemb object}
\usage{
\method{summary}{glemb}(object, digits = 3L, ...)
}
\arguments{
\item{object}{A \code{glemb} object returned by \code{\link[=glemb]{glemb()}}.}

\item{digits}{Integer. Number of decimal places for numeric values.
Default is \code{3}.}

\item{...}{Ignored.}
}
\value{
\code{object}, invisibly.
}
\description{
Prints a summary of the imputation model and the missingness in the first
imputed dataset.
}
